+++ src/components/Main.tsx
@@ -35,7 +35,7 @@ import { EdgeHeader } from './navigation/EdgeHeader'
 import { PluginBackButton } from './navigation/GuiPluginBackButton'
 import { HeaderBackground } from './navigation/HeaderBackground'
 import { HeaderTextButton } from './navigation/HeaderTextButton'
-import { ParamHeaderCryptoIconTitle, ParamHeaderTitle } from './navigation/ParamHeaderTitle'
+import { ParamHeaderTitle } from './navigation/ParamHeaderTitle'
 import { SideMenuButton } from './navigation/SideMenuButton'
 import { TransactionDetailsTitle } from './navigation/TransactionDetailsTitle'
 import { LoadingSplashScreen } from './progress-indicators/LoadingSplashScreen'
@@ -753,11 +753,7 @@ const EdgeWalletsTabScreen = () => {
       <Stack.Screen
         name="transactionList"
         component={TransactionList}
-        options={{
-          headerTitle: () => (
-            <ParamHeaderCryptoIconTitle<'transactionList'> fromCryptoIconParams={params => ({ walletId: params.walletId, tokenId: params.tokenId })} />
-          )
-        }}
+        options={{ headerTitle: () => <ParamHeaderTitle<'transactionList'> fromParams={params => params.walletName} /> }}
       />
     </Stack.Navigator>
   )