+++ src/constants/constantSettings.ts
@@ -1,6 +1,7 @@
 export const AAVE_WELCOME = 'aaveWelcome.json'
 export const FIRST_OPEN = 'firstOpen3.json'
-export const SCAM_WARNING = 'scamWarning.json'
+export const SCAM_WARNING = 'scamWarning.json' // For warnings on sends
+export const SCAM_WARNING_2 = 'scamWarning2.json' // For other general dangerous actions
 export const SETTINGS_PERMISSION_LIMITS = 'SETTINGS_PERMISSION_LIMIT'
 export const SETTINGS_PERMISSION_QUANTITY = 3
 export const LOCAL_EXPERIMENT_CONFIG = 'remoteConfigSticky.json'